Refactored prop/texture replacement
